// src/app/models/peticion.model.ts
export interface Peticion {
  id: string;
  id_usu_pet: string;
  nombre_usu: string;
  apellido_usu: string;
  fecha_pet: string;
  fecha_inicio: string;
  fecha_fin: string;
  hora_inicio: string;
  hora_fin: string;
  descripcion_pet: string;
  observacion_pet: string;
  estado_pet: string;
  id_tipo_pet_asig: string;
}